[fix] #323: Handle recipients pubkey with a checksum
Passed pubkeys with a checksum wasn't returned without the checksum

recipient/outputAddresses: Click stores the arguments in a tuple which
is immutable. Convert it to a list in order to be able to change it.
Though it is not necessary to return the value since it's a list.

outputAddresses: use enumerate in order to assign the pubkeys without the
checksum

ouputBackChange: remove duplicate call to check_public_key()
store pubkey in case it get erased by a boolean
Return the value as a string is not a pointer and does not get modified
in the parent scope as a list or a dict.

Add comments in check_transaction_values()

This feature was broken and forgotten since v0.5.0 release